+++ src/main/java/net/kyrptonaught/customportalapi/portal/PortalIgnitionSource.java
@@ -10,9 +10,15 @@ import net.minecraft.item.Item;
 import net.minecraft.tag.BlockTags;
 import net.minecraft.tag.FluidTags;
 import net.minecraft.util.Identifier;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.registry.Registry;
+import net.minecraft.world.World;
 
 import java.util.HashSet;
+import java.util.function.BiFunction;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.Supplier;
 
 public class PortalIgnitionSource {
     public final static PortalIgnitionSource FIRE = new PortalIgnitionSource(SourceType.BLOCKPLACED, Registry.BLOCK.getId(Blocks.FIRE));
@@ -44,6 +50,10 @@ public class PortalIgnitionSource {
 
     public static PortalIgnitionSource CustomSource(Identifier ignitionSourceID) {
         return new PortalIgnitionSource(SourceType.CUSTOM, ignitionSourceID);
+    }
+
+    public void withCondition(BiFunction<World, BlockPos, Boolean> condition) {
+        
     }
 
     public boolean isWater() {