+++ src/main/java/net/kyrptonaught/customportalapi/util/PortalLink.java
@@ -14,35 +14,19 @@ public class PortalLink {
     public Identifier dimID;
     public Identifier returnDimID = new Identifier("overworld");
     public int colorID;
+    public int forcedWidth, forcedHeight;
 
     public PortalLink(Identifier blockID, Identifier dimID, int colorID) {
         this.block = blockID;
         this.dimID = dimID;
         this.colorID = colorID;
-
-    }
-
-    public PortalLink(Identifier blockID, Identifier dimID, int r, int g, int b) {
-        this(blockID, dimID, ColorUtil.getColorFromRGB(r, g, b));
-    }
-
-    public PortalLink(Identifier blockID, PortalIgnitionSource portalIgnitionSource, Identifier dimID, int colorID) {
-        this(blockID, dimID, colorID);
-        this.portalIgnitionSource = portalIgnitionSource;
-    }
-
-    public PortalLink(Identifier blockID, PortalIgnitionSource portalIgnitionSource, CustomPortalBlock portalBlock, Identifier dimID, int colorID) {
-        this(blockID, portalIgnitionSource, dimID, colorID);
-        this.portalBlock = portalBlock;
-    }
-
-    public PortalLink(Identifier blockID, PortalIgnitionSource portalIgnitionSource, CustomPortalBlock portalBlock, Identifier dimID, Identifier returnDimId, int colorID) {
-        this(blockID, portalIgnitionSource, dimID, colorID);
-        this.portalBlock = portalBlock;
-        this.returnDimID = returnDimId;
     }
 
     public boolean doesIgnitionMatch(PortalIgnitionSource attemptedSource) {
         return portalIgnitionSource.sourceType == attemptedSource.sourceType && portalIgnitionSource.ignitionSourceID == attemptedSource.ignitionSourceID;
+    }
+
+    public boolean isCorrectForcedSize(int attempWidth, int attemptHeight) {
+        return ((forcedWidth == 0 || forcedWidth == attempWidth) && (forcedHeight == 0 || forcedHeight == attemptHeight));
     }
 }